Part-01 - Building a Discord Quiz Bot/requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
discord.py==1.7.1
Django==3.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
multidict==5.1.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
